requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
transformers==4.51.3
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.65.0